[BFC] Fix precision of model parameters
Before this commit the precision of model parameters was the one
of W_init, pi_init and sigma_init, not the one the model was configured
with.

With these changes, BSC supports initialization of model parameters
with tensors on any device and with any precision -- they are converted
to required device and precision internally.