Make sure all processes always loop over the same number of batches
Using ShufflingSampler as the torch sampler for TVEMDataLoader in
mpi runs we make sure that all processes run over the same number
of datapoints, and therefore the same number of batches.

This is desirable because it allows to simply do mpi scatter/gather
operations also within the processing of a batch -- a requirement,
for instance, for distributed stochastic gradient descent.