[TVAE] Fix gradient estimation
Before this patch, truncated posteriors were not considered constant
for the purposes of the evaluation of the free energy gradients, i.e.
the M-step was optimizing the parameters _and_ the truncated posteriors.

With this patch, we use a copy of the log-pseudo-joints that is detached
from the gradient computation in order to compute the mean over
posteriors of the log-joints, which results in a computation that
effectively considers truncated posteriors constant for the purposes
of the optimization of parameters in the M-step.

As a consequence TVAE is not able to evaluate partial values of the
free energy for every batch, so some code is updated to have
TVAE.update_update_param_batch return None instead of the free energy
value.